app/requirements.txt
django==4.0.4
channels==3.0.4
channels-redis==3.4.0
PyJWT
pydantic==1.9.1
